+++ FileDifferenceForensicEvidence/me.py
@@ -1,26 +1,22 @@
 import os
-import sys
-import logging
-import subprocess
-import time
+from shared_utilities import Configuration
 import shutil
 import re
-import shared_utilities
 
 def get_name():
     return "Merge evidences"
 
-def execute(configuration):
+def execute(configuration : Configuration):
     if os.path.exists(configuration.current_folder + "\\me\\"):
         shutil.rmtree(configuration.current_folder + "\\me\\")
     os.makedirs(configuration.current_folder + "\\me\\")
-    def write_content_merged(dictionary,output_me_file_with_full_path, label_for_operation):
+    def write_content_merged(dictionary:str,output_me_file_with_full_path:str, label_for_operation:str):
         result = ""
         for key in dictionary:
             result = result + key + "\t" + label_for_operation + "\t" + str(dictionary[key]) + "\n"
         with open(output_me_file_with_full_path, "a") as file:
             file.write(result)
-    def merge_evidence_for_file(input_pe_files_with_full_path,output_me_file_with_full_path):
+    def merge_evidence_for_file(input_pe_files_with_full_path, output_me_file_with_full_path:str):
         dictionary_access = {}
         dictionary_modify = {}
         dictionary_created = {}
@@ -60,7 +56,7 @@ def execute(configuration):
                             else:
                                 dictionary_deleted[found_file] = 1
 
-        if os.path.exists(output_me_file_with_full_path):
+        if os.path.exists(output_me_file_with_full_path and configuration.overwrite_existing_files_and_snapshots):
             os.remove(output_me_file_with_full_path)
         with open(output_me_file_with_full_path, 'w'):
             pass
@@ -79,9 +75,9 @@ def execute(configuration):
                 for i in range(1, configuration.amount_of_executions_per_action + 1):
                     pe_files.append(configuration.current_folder + "\\pe\\" + action[1] + "." + str(i) + ".pe")
                 merge_evidence_for_file(pe_files,configuration.current_folder + "\\me\\" + action[1] + ".me")
-            except Exception as exception:
+            except Exception as exception_argument:
                 configuration.log.error("Exception occurred while merge evidence  for action " + action[1] + ":")
-                configuration.log.error(exception, exc_info=True)
+                configuration.log.error(exception_argument, exc_info=True)
             configuration.log.info("Merge evidence for action " + action[1] + " finished")
             
     try:
